test: migrate form field tests to form_fields.schema.spec
